separate component for each comment
